+++ src/test/java/blackjack/domain/GameBoardTest.java
@@ -6,6 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import blackjack.domain.card.Deck;
 import java.util.List;
 import java.util.Map;
+
+import blackjack.domain.card.TestDeckFactory;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Gamers;
 import blackjack.domain.gamer.Player;
