+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -2,7 +2,6 @@ package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.TestDeckFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
