+++ src/test/java/blackjack/domain/card/TestDeckFactory.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.DeckFactory;
